Tenzing/CA Auto restart, Misc other changes

+++ Scripts/Dev/mr_test.sh
@@ -11,12 +11,18 @@ cd ${mr_client_src_dir}/
 
 print "Running test ..."
 
-hdfs dfs -mkdir /test
-hdfs dfs -copyFromLocal ${sherpa_src_dir}/jobSubPlusCounters/core/src/main/java/com/sherpa/core/dao/WorkloadCountersPhoenixDAO.java /test/
 
-#yarn jar ${hadoop_home}/share/hadoop/mapreduce/hadoop-mapreduce-examples-2.6.0.jar pi 10 100
+if [ $# -eq 1 ]
+  then
+    hdfs dfs -rm -r /mrTestOutputBySherpa
+    hdfs dfs -mkdir /test
+    hdfs dfs -copyFromLocal ${sherpa_src_dir}/jobSubPlusCounters/core/src/main/java/com/sherpa/core/dao/WorkloadCountersPhoenixDAO.java /test/
+    yarn jar ${hadoop_home}/share/hadoop/mapreduce/hadoop-mapreduce-examples-${HADOOP_VERSION}.jar wordcount -D PSManaged=true -D SherpaCostObj=Latency /test/ /mrTestOutputBySherpa
+  else
+    yarn jar ${hadoop_home}/share/hadoop/mapreduce/hadoop-mapreduce-examples-${HADOOP_VERSION}.jar pi -D PSManaged=true -D SherpaCostObj=Latency 5 5
+fi
 
 
-yarn jar ${hadoop_home}/share/hadoop/mapreduce/hadoop-mapreduce-examples-${HADOOP_VERSION}.jar wordcount -D PSManaged=true -D SherpaCostObj=Latency /test/ /mrTestOutputBySherpa
-hdfs dfs -rm -r /mrTestOutputBySherpa
+
+
